test: split dump reload tests
