+++ tests/distributed/test_remote_flow_dump_rolling_update_stress/flow_dbms.yml
@@ -1,0 +1,9 @@
+jtype: Flow
+version: '1'
+with:
+  read_only: false
+  rest_api: true
+  port_expose: 9000
+pods:
+  - name: indexer_dbms
+    uses: indexer_dbms.yml
