+++ tests/distributed/test_remote_flow_dump_rolling_update_stress/flow_query.yml
@@ -1,0 +1,14 @@
+jtype: Flow
+version: '1'
+with:
+  read_only: false
+  rest_api: true
+  port_expose: 9001
+pods:
+  - name: indexer_query
+    uses: indexer_query.yml
+    replicas: 2
+    parallel: 3
+    polling: all
+    uses_after: _merge_matches
+    timeout_ready: '-1'
